requirements.txt
streamlit
pydantic-ai
httpx
logfire
python-dotenv
devtools
